Version 0.3.0. Add input validation and error handling to prevent crashes. Now supports Node.js 16-23.
- Added argument count and type validation in C++ addon methods
- Added null result checking for sha256_crypt/sha512_crypt failures
- Added error handling in async worker classes with SetErrorMessage
- Fixed Makefile to use local mocha installation
- Added comprehensive validation test cases

+++ src/sha256crypt.c
@@ -449,7 +449,7 @@ sha256_crypt_r (const char *key, const char *salt, char *buffer, int buflen)
   sha256_init_ctx (&alt_ctx);
 
   /* For every character in the password add the entire password.  */
-  for (cnt = 0; cnt < 16 + alt_result[0]; ++cnt)
+  for (cnt = 0; cnt < (size_t)(16 + alt_result[0]); ++cnt)
     sha256_process_bytes (salt, salt_len, &alt_ctx);
 
   /* Finish the digest.  */
